requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-docx
faiss-cpu
chromadb
sentence-transformers
langchain
langchain-groq
python-dotenv
pdfplumber
pymongo
rank-bm25